Move viewmodels to viewmodels package
