+++ feature/shoppinglist/src/main/java/co/ke/xently/shoppinglist/ShoppingListActivity.kt
@@ -27,7 +27,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
 import co.ke.xently.data.ShoppingListItem
-import co.ke.xently.feature.LocationPermissionViewModel
+import co.ke.xently.feature.viewmodels.LocationPermissionViewModel
 import co.ke.xently.feature.LocationService
 import co.ke.xently.feature.theme.XentlyTheme
 import co.ke.xently.feature.ui.NavMenuItem
