+++ feature/shoppinglist/src/main/java/co/ke/xently/shoppinglist/ui/list/recommendation/ShoppingListRecommendationViewModel.kt
@@ -5,7 +5,7 @@ import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.viewModelScope
 import co.ke.xently.data.RecommendationReport
 import co.ke.xently.data.TaskResult
-import co.ke.xently.feature.LocationPermissionViewModel
+import co.ke.xently.feature.viewmodels.LocationPermissionViewModel
 import co.ke.xently.feature.utils.flagLoadingOnStart
 import co.ke.xently.shoppinglist.Recommend
 import co.ke.xently.shoppinglist.repository.IShoppingListRepository
